lambda/requirements.txt
numpy==1.26.4
onnxruntime==1.16.3
pillow==10.4.0
